+++ docs/CHANGELOG.rst
@@ -9,6 +9,11 @@ All notable changes to this project are documented in this file.
 Unreleased
 ==========
 
+Changed
+-------
+* **BACKWARD INCOMPATIBLE:** Change parameter ``email`` to ``username``
+  in Resolwe constructor
+
 
 ==================
 3.0.0 - 2018-02-21